# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SARTU API",
        "description": "Documentación OpenAPI mínima",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/empresa": {
            "get": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Listado de empresas (vista)",
                "operationId": "8913427cbfbc6b6759d466ef71e6e1a2",
                "responses": {
                    "200": {
                        "description": "HTML view"
                    }
                }
            },
            "post": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Crear empresa (form submission)",
                "operationId": "921673159512f786566f64b031d6bd36",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "limite_usuarios": {
                                        "type": "integer"
                                    },
                                    "jornada_diaria_minutos": {
                                        "type": "integer"
                                    },
                                    "max_pausa_no_contabilizada": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect (HTML)"
                    }
                }
            }
        },
...
        "/empresa/{empresa}/edit": {
            "get": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Editar empresa (formulario)",
                "operationId": "42d39b64874512c6c3720733740e8dc3",
                "parameters": [
                    {
                        "name": "empresa",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTML view"
                    }
                }
            }
        },
        "/empresa/{empresa}": {
            "put": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Actualizar empresa (form submission)",
                "operationId": "070c9b4a0878e4de79835e8eda3a88e2",
                "parameters": [
                    {
                        "name": "empresa",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "limite_usuarios": {
                                        "type": "integer"
                                    },
                                    "jornada_diaria_minutos": {
                                        "type": "integer"
                                    },
                                    "max_pausa_no_contabilizada": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect (HTML)"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Eliminar empresa",
                "operationId": "b32af6e9f9a08e8ffcd503d22c24cda0",
                "parameters": [
                    {
                        "name": "empresa",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect (HTML)"
                    }
                }
            }
        },
...
        "/fichajes": {
            "get": {
                "tags": [
                    "Fichaje"
                ],
                "summary": "Resumen general (encargado/admin)",
                "operationId": "cb3fbe10b85205f60eea0431b4fc387b",
                "parameters": [
                    {
                        "name": "fecha",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTML view"
                    }
                }
            },
            "post": {
                "tags": [
                    "Fichaje"
                ],
                "summary": "Registrar fichaje (form submission)",
                "operationId": "c044069f6f6ce03589a5d3b4b0a5f5ac",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "tipo": {
                                        "type": "string",
                                        "example": "entrada"
                                    },
                                    "metodo": {
                                        "type": "string",
                                        "example": "web"
                                    },
                                    "lat": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "lng": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "notas": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect back with success"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/fichajes/resumen-empresa": {
            "get": {
                "tags": [
                    "Fichaje"
                ],
                "summary": "Resumen de empresa por fecha",
                "operationId": "4672ee2327416b3309f36845a75a169b",
                "parameters": [
                    {
                        "name": "fecha",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "empresa_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTML view"
                    }
                }
            }
        },
        "/fichajes/estado-empresa": {
            "get": {
                "tags": [
                    "Fichaje"
                ],
                "summary": "Estado actual de empleados por empresa",
                "operationId": "35baa386e1b371a854e2ed16eaf8e763",
                "parameters": [
                    {
                        "name": "empresa_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTML view"
                    }
                }
            }
        },
...
    },
...
}